refactor(decidables-elements): add subfolders to `src`
Create `components`, `mixins`, and `utilities` subfolders

BREAKING CHANGE: subfolders change `import` paths
