+++ libraries/decidables-elements/src/components/index.js
@@ -1,0 +1,7 @@
+// UI Elements
+export {default as DecidablesButton} from './button';
+export {default as DecidablesSlider} from './slider';
+export {default as DecidablesSpinner} from './spinner';
+export {default as DecidablesSwitch} from './switch';
+export {default as DecidablesToggle} from './toggle';
+export {default as DecidablesToggleOption} from './toggle-option';
