+++ libraries/decidables-elements/src/index.js
@@ -1,19 +1,6 @@
 
 // Base class for Decidables Web Components
 export {default as DecidablesElement} from './decidables-element';
-
-// UI Elements
-export {default as DecidablesButton} from './button';
-export {default as DecidablesSlider} from './slider';
-export {default as DecidablesSpinner} from './spinner';
-export {default as DecidablesSwitch} from './switch';
-export {default as DecidablesToggle} from './toggle';
-export {default as DecidablesToggleOption} from './toggle-option';
-
-// Property Converters
-export {default as DecidablesConverterArray} from './converter-array';
-export {default as DecidablesConverterSet} from './converter-set';
-
-// Mixins
-export {default as DecidablesMixinEquation} from './mixin-equation';
-export {default as DecidablesMixinResizeable} from './mixin-resizeable';
+export * from './components';
+export * from './mixins';
+export * from './utilities';
