+++ libraries/decidables-elements/src/mixins/index.js
@@ -1,0 +1,3 @@
+// Mixins
+export {default as DecidablesMixinEquation} from './mixin-equation';
+export {default as DecidablesMixinResizeable} from './mixin-resizeable';
