+++ libraries/decidables-elements/src/mixins/mixin-equation.js
