+++ libraries/decidables-elements/src/mixins/mixin-resizeable.js
