+++ libraries/decidables-elements/src/utilities/converter-array.js
